fix(cli-sub-agent): validate model-spec against config
Validate explicit model specs against configured tiers or tool defaults so unconfigured specs are rejected before execution. Allow explicit --tool to filter a selected tier while preserving tier model/thinking routing, and keep --force-ignore-tier-setting as the bypass. Split run helper internals to satisfy monolith gates and bump workspace version to 0.1.794.
